requirements.txt
django==5.1.4
Faker==26.0.0
crispy-bootstrap5==2024.2.0
django_widget_tweaks==1.5.0
django-crispy-forms==2.3.0
django_bootstrap5==24.2.0
pandas==2.2.2
gunicorn==22.0.0
django-heroku==0.3.1
plotly==5.22.0
openpyxl
django-star-ratings